\name{cm.ci}
\alias{cm.ci}
\title{Cousineau-Morey Within-Subject CIs}
\description{Calcualte Cousineau-Morey within-subject CIs}
\usage{cm.ci(data.frame, conf.level = 0.95, difference = TRUE)}
\arguments{
  \item{data.frame}{Data frame containing input to calculate confidence intervals on}
  \item{conf.level}{Significance level, alpha}
  \item{difference}{blah}
}
\value{Matrix containing confidence intervals}
\author{Timothy R. Koscik <timkoscik+rmisc@gmail.com>}