\name{decompose.var}
\alias{decompose.var}
\title{temporary title}
\description{}
\usage{
decompose.var(df, var.name, grouping=c("grand", "subject", "subject.run")) 
}
\arguments{
  \item{df}{}
  \item{var.name}{}
  \item{grouping}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}