\name{faster.cor}
\alias{faster.cor}
\title{Faster Pearson Correlation}
\description{A faster implementation for calculating pearson correlation coefficients}
\usage{faster.cor(inputMx, method = "pearson")}
\arguments{
  \item{inputMx}{m x n input matrix}
  \item{method}{Unused}
}
\value{n x n matrix of correlation coefficients}
\author{Timothy R. Koscik <timkoscik+rmisc@gmail.com>}