\name{good.color}
\alias{good.color}
\title{Color palettes optimized for colorblindness and grayscale printing}
\description{}
\usage{good.color(type = c("cool", "warm", "neon", "neon-bright"), n=21)}
\arguments{
  \item{type}{c("cool", "warm", "neon", "neon-bright")}
  \item{n}{number of colors to return}
}
\value{a character vector of color values in hex}
\author{Timothy R. Koscik <timkoscik+rmisc@gmail.com>}
