\name{outliers.by.std.resid}
\alias{outliers.by.std.rsid}
\title{Remove outliers from LMER/BRM/LM models by standardized residual}
\description{}
\usage{
outliers.by.std.resid(model, data, trim = 2.5)
}
\arguments{
  \item{model}{}
  \item{data}{}
  \item{trim}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}