\name{outliers.by.subject}
\alias{outliers.by.subject}
\title{Remove outliers from LMER/BRM/LM models by standardized residual}
\description{}
\usage{
outliers.by.subject(data, response, subject, trim = 2.5)
}
\arguments{
  \item{data}{}
  \item{response}{}
  \item{subject}{}
  \item{trim}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}