\name{quantile.scale}
\alias{quantile.scale}
\title{temporary title}
\description{}
\usage{
quantile.scale(df, var.name, center=TRUE, prob=0.9)
}
\arguments{
  \item{df}{}
  \item{var.name}{}
  \item{center}{}
  \item{prob}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}