\name{tkmisc-package}
\alias{tkmisc-package}
\alias{tkmisc}
\docType{package}
\title{
\packageTitle{tkmisc}
}
\description{
\packageDescription{tkmisc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tkmisc}
\packageIndices{tkmisc}
TKoscik's Miscellaneous Functions
}
\author{
\packageAuthor{tkmisc}

Maintainer: \packageMaintainer{tkmisc}
}
