\name{tk.r.misc-package}
\alias{tk.r.misc-package}
\alias{tk.r.misc}
\docType{package}
\title{
\packageTitle{tk.r.misc}
}
\description{
\packageDescription{tk.r.misc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tk.r.misc}
\packageIndices{tk.r.misc}
TKoscik's Miscellaneous Functions
}
\author{
\packageAuthor{tk.r.misc}

Maintainer: \packageMaintainer{tk.r.misc}
}